UnityXRCameraRig
Added a UnityXRCameraRig scene so that I could get interactions working for the Vive so that I could help with interactions within VR also.  Was able to make it so that I can pick up and throw books / key using trigger as well.  Messed around with hand collisions but nothing concrete yet.